🎨 Feat: basic.test.tsx
